Initial commit of workspace management
Includes script that can be executed against either a Terraform Enterprise or Terraform Cloud deployment, using a Bearer token for authentication.

Can accept a list of orgs to operate on, or no org if all orgs are to be operated on.

workspaces command will return a list of terraform resources for workspaces.